refactor: separate node.js logic from shared webpack compat
