+++ packages/pack/src/commands/build.ts
@@ -5,6 +5,7 @@ import { nanoid } from "nanoid";
 import { join } from "path";
 import {
   compatOptionsFromWebpack,
+  readWebpackConfig,
   WebpackConfig,
 } from "../config/webpackCompat";
 import { projectFactory } from "../core/project";
@@ -20,9 +21,18 @@ export function build(
   projectPath?: string,
   rootPath?: string,
 ) {
-  const bundleOptions = (<WebpackConfig>options).webpackMode
-    ? compatOptionsFromWebpack(<WebpackConfig>options, projectPath, rootPath)
-    : <BundleOptions>options;
+  let bundleOptions: BundleOptions;
+  if ((<WebpackConfig>options).webpackMode) {
+    let webpackConfig = <WebpackConfig>options;
+    if (!webpackConfig.entry) {
+      const loadedConfig = readWebpackConfig(projectPath, rootPath);
+      webpackConfig = { ...webpackConfig, ...loadedConfig };
+    }
+    bundleOptions = compatOptionsFromWebpack(webpackConfig);
+  } else {
+    bundleOptions = <BundleOptions>options;
+  }
+
   if (!rootPath) {
     // help user to find the rootDir automatically.
     rootPath = findRootDir(projectPath || process.cwd());
