+++ packages/utoo-web/src/index.ts
@@ -1,6 +1,5 @@
 export {
   compatOptionsFromWebpack,
-  readWebpackConfig,
   type WebpackConfig,
 } from "@utoo/pack-shared";
 export { Project } from "./project";
